pyproject.toml
[tool.poetry]
name = "oglaf-feed-bot"
version = "0.1.0"
description = ""
authors = ["Nyaaa <1457047+Nyaaa@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "oglaf_feed_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.23.1"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
